feat: create ui module
